+++ test/libsolidity/syntaxTests/functionTypes/function_type_return_parameters_with_names.sol
@@ -1,0 +1,5 @@
+contract C {
+    function(uint) returns (bool ret) f;
+}
+// ----
+// SyntaxError: (41-49): Return parameters in function types may not be named.