+++ test/libsolidity/syntaxTests/parsing/function_type_in_struct.sol
@@ -1,6 +1,6 @@
 contract test {
     struct S {
-        function (uint x, uint y) internal returns (uint a) f;
+        function (uint x, uint y) internal returns (uint) f;
         function (uint, uint) external returns (uint) g;
         uint d;
     }
@@ -8,4 +8,3 @@ contract test {
 // ----
 // Warning: (49-55): Naming function type parameters is deprecated.
 // Warning: (57-63): Naming function type parameters is deprecated.
-// Warning: (83-89): Naming function type return parameters is deprecated.